Repo Description
Examples of loopers, handlers, & threads.
ProgressBarSimulatedFileDownload.zip contains a project that simulates a
file download and updates the progress bar.

ProgressBarNetworkFileDownload.zip contains a project that downloads a
real image file and displays it in an ImageView. It also uses a Looper
and a Handler to update a ProgressDialog, depending on how many bytes
have been downloaded.